requirements.txt
pytesseract
opencv-python
pdf2image
Pillow
pandas
google-generativeai